+++ velocityfield/src/velfntest.cpp
@@ -10,9 +10,9 @@ int main(int i_iArg_Count, const char * i_lpszArg_Values[])
 	Bose_Einstein_velocity_function	vfBE(10000.0);
 
 	printf("Vel, Max, Fermi, BE\n");
-	for (double dVel = 1e-10; dVel < 1e10; dVel *= 1.25)
+	for (long double dVel = 1e-10; dVel < 1e10; dVel *= 1.25)
 	{
-		printf("%.2e, %.2e, %.2e, %.2e\n",dVel,vfMax(dVel),vfFermi(dVel),vfBE(dVel));
+		printf("%.2Le, %.2Le, %.2Le, %.2Le\n",dVel,vfMax(dVel),vfFermi(dVel),vfBE(dVel));
 	}
 	
 	return 0;
